requirements.txt
aioredis==1.2.0
asgiref==2.3.2
async-timeout==2.0.1
attrs==18.2.0
autobahn==18.10.1
Automat==0.7.0
channels==2.1.5
channels-redis==2.3.1
constantly==15.1.0
daphne==2.2.2
Django==2.1.2
django-rest-framework==0.1.0
djangorestframework==3.9.0
hiredis==0.2.0
hyperlink==18.0.0
idna==2.7
incremental==17.5.0
msgpack==0.5.6
PyHamcrest==1.9.0
pytz==2018.6
six==1.11.0
Twisted==18.9.0
txaio==18.8.1
zope.interface==4.6.0
